requirements.txt
Django==4.2.4
django-crispy-forms==2.0
Pillow==10.0.0
numpy==1.23.5
opencv-python==4.7.0.72
mtcnn==0.1.1
tensorflow==2.12.0
scikit-learn==1.2.2
matplotlib==3.7.1
scipy==1.10.1
python-dotenv==1.0.0
gunicorn==20.1.0
django-tables2==2.6.0
django-filter==23.2
django-widget-tweaks==1.5.0